add simplier way to configure payload generators
