+++ corehq/apps/repeaters/models.py
@@ -1,34 +1,32 @@
 from collections import namedtuple
 from datetime import datetime, timedelta
-import logging
 import urllib
 import urlparse
 import warnings
+
 from django.utils.translation import ugettext_lazy as _
 from requests.auth import HTTPBasicAuth, HTTPDigestAuth
-
 from requests.exceptions import Timeout, ConnectionError
+from couchdbkit.exceptions import ResourceNotFound
+
 from corehq.apps.cachehq.mixins import QuickCachedDocumentMixin
 from corehq.apps.locations.models import SQLLocation
+from corehq.apps.repeaters.repeater_generators import FormRepeaterXMLPayloadGenerator, \
+    FormRepeaterJsonPayloadGenerator, CaseRepeaterXMLPayloadGenerator, CaseRepeaterJsonPayloadGenerator, \
+    ShortFormRepeaterJsonPayloadGenerator, AppStructureGenerator, UserPayloadGenerator, LocationPayloadGenerator
 from corehq.apps.users.models import CommCareUser
 from corehq.form_processor.exceptions import XFormNotFound
 from corehq.util.datadog.metrics import REPEATER_ERROR_COUNT
 from corehq.util.datadog.gauges import datadog_counter
 from corehq.util.quickcache import quickcache
-
 from dimagi.ext.couchdbkit import *
-from couchdbkit.exceptions import ResourceNotFound
-
 from casexml.apps.case.xml import V2, LEGAL_VERSIONS
-from corehq.apps.receiverwrapper.exceptions import DuplicateFormatException
 from corehq.form_processor.interfaces.dbaccessors import FormAccessors, CaseAccessors
-
 from couchforms.const import DEVICE_LOG_XMLNS
 from dimagi.utils.decorators.memoized import memoized
 from dimagi.utils.parsing import json_format_datetime
 from dimagi.utils.mixins import UnicodeMixIn
 from dimagi.utils.post import simple_post
-
 from .dbaccessors import (
     get_pending_repeat_record_count,
     get_failure_repeat_record_count,
@@ -53,67 +51,6 @@ def log_repeater_timeout_in_datadog(domain):
 
 
 DELETED = "-Deleted"
-
-FormatInfo = namedtuple('FormatInfo', 'name label generator_class')
-
-
-class GeneratorCollection(object):
-    """Collection of format_name to Payload Generators for a Repeater class
-
-    args:
-        repeater_class: A valid child class of Repeater class
-    """
-
-    def __init__(self, repeater_class):
-        self.repeater_class = repeater_class
-        self.default_format = ''
-        self.format_generator_map = {}
-
-    def add_new_format(self, format_name, format_label, generator_class, is_default=False):
-        """Adds a new format->generator mapping to the collection
-
-        args:
-            format_name: unique name to identify the format
-            format_label: label to be displayed to the user
-            generator_class: child class of .repeater_generators.BasePayloadGenerator
-
-        kwargs:
-            is_default: True if the format_name should be default format
-
-        exceptions:
-            raises DuplicateFormatException if format is added with is_default while other
-            default exists
-            raises DuplicateFormatException if format_name alread exists in the collection
-        """
-        if is_default and self.default_format:
-            raise DuplicateFormatException("A default format already exists for this repeater.")
-        elif is_default:
-            self.default_format = format_name
-        if format_name in self.format_generator_map:
-            raise DuplicateFormatException("There is already a Generator with this format name.")
-
-        self.format_generator_map[format_name] = FormatInfo(
-            name=format_name,
-            label=format_label,
-            generator_class=generator_class
-        )
-
-    def get_default_format(self):
-        """returns default format"""
-        return self.default_format
-
-    def get_default_generator(self):
-        """returns generator class for the default format"""
-        raise self.format_generator_map[self.default_format].generator_class
-
-    def get_all_formats(self, for_domain=None):
-        """returns all the formats added to this repeater collection"""
-        return [(name, format.label) for name, format in self.format_generator_map.iteritems()
-                if not for_domain or format.generator_class.enabled_for_domain(for_domain)]
-
-    def get_generator_by_format(self, format):
-        """returns generator class given a format"""
-        return self.format_generator_map[format].generator_class
 
 
 class Repeater(QuickCachedDocumentMixin, Document, UnicodeMixIn):
@@ -324,6 +261,13 @@ class FormRepeater(Repeater):
 
     """
 
+    class Formats(object):
+        formats = {
+            'form_xml': (FormRepeaterXMLPayloadGenerator, _('XML')),
+            'form_json': (FormRepeaterJsonPayloadGenerator, _('JSON')),
+        }
+        default_format = 'form_xml'
+
     include_app_id_param = BooleanProperty(default=True)
     white_listed_form_xmlns = StringListProperty(default=[])  # empty value means all form xmlns are accepted
     friendly_name = _("Forward Forms")
@@ -370,6 +314,13 @@ class CaseRepeater(Repeater):
 
     """
 
+    class Formats(object):
+        formats = {
+            'case_xml': (CaseRepeaterXMLPayloadGenerator, _('XML')),
+            'case_json': (CaseRepeaterJsonPayloadGenerator, _('JSON')),
+        }
+        default_format = 'case_xml'
+
     version = StringProperty(default=V2, choices=LEGAL_VERSIONS)
     white_listed_case_types = StringListProperty(default=[])  # empty value means all case-types are accepted
     black_listed_users = StringListProperty(default=[])  # users who caseblock submissions should be ignored
@@ -409,6 +360,12 @@ class ShortFormRepeater(Repeater):
 
     """
 
+    class Formats(object):
+        formats = {
+            'short_form_json': (ShortFormRepeaterJsonPayloadGenerator, _('Default JSON')),
+        }
+        default_format = 'short_form_json'
+
     version = StringProperty(default=V2, choices=LEGAL_VERSIONS)
     friendly_name = _("Forward Form Stubs")
 
@@ -433,12 +390,24 @@ class ShortFormRepeater(Repeater):
 class AppStructureRepeater(Repeater):
     friendly_name = _("Forward App Schema Changes")
 
+    class Formats(object):
+        formats = {
+            'id': (AppStructureGenerator, _('ID Only')),
+        }
+        default_format = 'app_structure_id'
+
     def payload_doc(self, repeat_record):
         return None
 
 
 class UserRepeater(Repeater):
     friendly_name = _("Forward Users")
+
+    class Formats(object):
+        formats = {
+            'user_json': (UserPayloadGenerator, _("JSON")),
+        }
+        default_format = 'user_json'
 
     @memoized
     def payload_doc(self, repeat_record):
@@ -450,6 +419,12 @@ class UserRepeater(Repeater):
 
 class LocationRepeater(Repeater):
     friendly_name = _("Forward Locations")
+
+    class Formats(object):
+        formats = {
+            'location_json': (LocationPayloadGenerator, _("JSON")),
+        }
+        default_format = 'location_json'
 
     @memoized
     def payload_doc(self, repeat_record):
